requirements.txt
Django==1.11.6
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
djangorestframework==3.7.0
pytz==2017.2
sqlparse==0.2.4
